Restructured API layourt
